requirements.txt
# SPDX-FileCopyrightText: Copyright 2022 EDF (Électricité de France S.A.)
# SPDX-License-Identifier: BSD-3-Clause
# See README for all details on copyright, authorship and license.
pycose == 1.0.1
cbor2 == 5.4.6
pyyaml
